requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
pandas==2.2.2
pyyaml==6.0.2
requests==2.32.3
requests-cache==1.2.0
trafilatura==1.7.0
python-dateutil==2.9.0.post0
pytest==8.3.2
